selftests: breakpoint_test: ptrace poke targets should be long ints
ptrace pokes and peeks work at a word granularity, so nr_tests and
trapped need to be long ints.  Otherwise the PTRACE_POKEDATA calls will
trash an adjacent half-word on x86_64.

This fixes test failures when building with clang >= 3.7 at all
optimization levels, including -O0.  clang packs nr_tests and trapped
into the same word, causing the ptrace(PTRACE_POKEDATA, &trapped) call
in check_success() to trash the child's nr_tests.

+++ tools/testing/selftests/breakpoints/breakpoint_test.c
@@ -34,7 +34,7 @@ static pid_t child_pid;
  * the same test sequence. (ie: that we haven't forgotten
  * to call check_trapped() somewhere).
  */
-static int nr_tests;
+static long nr_tests;
 
 static void set_breakpoint_addr(void *addr, int n)
 {
@@ -127,7 +127,7 @@ static void (*dummy_funcs[])(void) = {
 	dummy_func3,
 };
 
-static int trapped;
+static long trapped;
 
 static void check_trapped(void)
 {
@@ -262,7 +262,7 @@ static void trigger_tests(void)
 static void check_success(const char *msg)
 {
 	const char *msg2;
-	int child_nr_tests;
+	long child_nr_tests;
 	int status;
 
 	/* Wait for the child to SIGTRAP */
